+++ unity-plugin/Editor/RegionTool/Drawing/PointByPointMode.cs
@@ -84,7 +84,7 @@ namespace UnityMCP.Editor.RegionTool
             return false;
         }
 
-        public Polygon2D? Finalize()
+        public Polygon2D? Build()
         {
             if (_vertices.Count < 3) return null;
             return new Polygon2D(_vertices.ToArray());